drivers: counter: stm32: introduce counter capture api
This intoduces apis for input capture on the STM32. This is something
agnostic enough that is seen on other vendors such as Atmel's SAM line
and others.

A use case for this would be timestamping a rising and/or falling edge
where percision and lack of jitter is important. The only way to currently
do this in zephyr is to set up a GPIO interrupt and then read the counter
or kernel time... but this can have limitations of software such as the
processor may have a lot going on and can be subject to interrupt latencies.
Having a hardware 'latch' done of the rising and/or falling edge with
respect to a timer can eliminate these impercisions.

How this is to be used, is that an application is to call
`capture_callback_set` with the configuration of the channel number, flags
(such as falling, rising, or both edges), and then a callback function. Then
the `capture_enable` function is to be called to enable this. The callback
has the timestamp of the edge in ticks and the edge that it captured (rising
or falling).

Signed-off-by: Ryan McClelland <ryanmcclelland@meta.com>