drivers: counter: change freq to 64 bits
Some counters, such as those that use fractional adders, tick at resolutions
greater than 4294967295Hz. This changes the freq in the common info
to a uint64_t. This retains the existing `counter_get_freq` function which
will still return a 32bit value. This also adds a `counter_get_freq_64`
function to get a 64bit value for the frequency.

Signed-off-by: Ryan McClelland <ryanmcclelland@meta.com>